+++ tiled/mimetypes.py
@@ -19,6 +19,12 @@ DEFAULT_ADAPTERS_BY_MIMETYPE = OneShotCachedMap(
         "multipart/related;type=image/tiff": lambda: importlib.import_module(
             "..adapters.tiff", __name__
         ).TiffSequenceAdapter.from_uris,
+        "image/jpeg": lambda: importlib.import_module(
+            "..adapters.jpeg", __name__
+        ).JPEGAdapter,
+        "multipart/related;type=image/jpeg": lambda: importlib.import_module(
+            "..adapters.jpeg", __name__
+        ).JPEGSequenceAdapter.from_uris,
         "text/csv": lambda: importlib.import_module(
             "..adapters.csv", __name__
         ).CSVAdapter,